Helper methods for loading 'kartverkets' placename data into an solr database.
Summary:
Helper classes and methods to parse 'karteverkets' placename in geojson and
to load them into a solr instance.

The patch also contains the solr schema.

Test for loading into solr is missing. I have not figured out
how we can write test for this without setting up a solr server. There
should be posible to run a solr standalone server as part of running the
tests but I need to find out how to do it.

The classes could also be better documented.

Test Plan:
Not much to test yet. But there is tests for the json parser.

activator clean test

Reviewers: #bora-dev

Differential Revision: https://phab.met.no/D737